+++ lib/nodejieba.cpp
@@ -9,6 +9,7 @@ cppjieba::KeywordExtractor* global_extractor_handle;
 NAN_METHOD(load) {
   if(info.Length() != 5) {
     info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
   }
 
   String::Utf8Value dictPath(info[0]->ToString());
@@ -37,51 +38,109 @@ NAN_METHOD(insertWord) {
     string word = *(String::Utf8Value(info[i]->ToString()));
     if(!global_jieba_handle->InsertUserWord(word)) {
       info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+      return;
     }
   }
   info.GetReturnValue().Set(Nan::New<v8::Boolean>(true));
-  
 }
+
+//NAN_METHOD(cut) {
+//  if (info.Length() == 0) {
+//    info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+//    return;
+//  }
+//  string sentence = *(String::Utf8Value(info[0]->ToString()));
+//  vector<string> words;
+//
+//  assert(global_jieba_handle);
+//  if (info.Length() >= 2) {
+//    string method = *(String::Utf8Value(info[1]->ToString()));
+//    if ("MP" == method) {
+//      global_jieba_handle->Cut(sentence, words, false); 
+//      if (info.Length() == 3) {
+//        global_jieba_handle->CutSmall(sentence, words, info[2]->Int32Value()); 
+//      }
+//    } else if ("HMM" == method) {
+//      global_jieba_handle->CutHMM(sentence, words); 
+//    } else if ("MIX" == method) {
+//      global_jieba_handle->Cut(sentence, words, true); 
+//    } else {
+//      global_jieba_handle->Cut(sentence, words, true); 
+//    }
+//  } else {
+//    global_jieba_handle->Cut(sentence, words, true); 
+//  }
+//
+//  Local<Array> outArray;
+//  WrapVector(words, outArray);
+//
+//  info.GetReturnValue().Set(outArray);
+//}
 
 NAN_METHOD(cut) {
   if (info.Length() == 0) {
     info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
   }
   string sentence = *(String::Utf8Value(info[0]->ToString()));
   vector<string> words;
-
-  assert(global_jieba_handle);
-  if (info.Length() >= 2) {
-    string method = *(String::Utf8Value(info[1]->ToString()));
-    if ("MP" == method) {
-      global_jieba_handle->Cut(sentence, words, false); 
-      if (info.Length() == 3) {
-        global_jieba_handle->CutSmall(sentence, words, info[2]->Int32Value()); 
-      }
-    } else if ("HMM" == method) {
-      global_jieba_handle->CutHMM(sentence, words); 
-    } else if ("MIX" == method) {
-      global_jieba_handle->Cut(sentence, words, true); 
-    } else if ("FULL" == method) {
-      global_jieba_handle->CutAll(sentence, words); 
-    } else if ("QUERY" == method) {
-      global_jieba_handle->CutForSearch(sentence, words, true); 
-    } else {
-      global_jieba_handle->Cut(sentence, words, true); 
-    }
-  } else {
-    global_jieba_handle->Cut(sentence, words, true); 
+  bool useHMM = false;
+  if (info.Length() > 1) {
+    useHMM = *info[1]->ToBoolean();
   }
-
+  global_jieba_handle->Cut(sentence, words, useHMM); 
   Local<Array> outArray;
   WrapVector(words, outArray);
+  info.GetReturnValue().Set(outArray);
+}
 
+NAN_METHOD(cutHMM) {
+  if (info.Length() == 0) {
+    info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
+  }
+  string sentence = *(String::Utf8Value(info[0]->ToString()));
+  vector<string> words;
+  global_jieba_handle->CutHMM(sentence, words); 
+  Local<Array> outArray;
+  WrapVector(words, outArray);
+  info.GetReturnValue().Set(outArray);
+}
+
+NAN_METHOD(cutAll) {
+  if (info.Length() == 0) {
+    info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
+  }
+  string sentence = *(String::Utf8Value(info[0]->ToString()));
+  vector<string> words;
+  global_jieba_handle->CutAll(sentence, words); 
+  Local<Array> outArray;
+  WrapVector(words, outArray);
+  info.GetReturnValue().Set(outArray);
+}
+
+NAN_METHOD(cutForSearch) {
+  if (info.Length() == 0) {
+    info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
+  }
+  string sentence = *(String::Utf8Value(info[0]->ToString()));
+  vector<string> words;
+  bool useHMM = false;
+  if (info.Length() > 1) {
+    useHMM = *info[1]->ToBoolean();
+  }
+  global_jieba_handle->CutForSearch(sentence, words, useHMM); 
+  Local<Array> outArray;
+  WrapVector(words, outArray);
   info.GetReturnValue().Set(outArray);
 }
 
 NAN_METHOD(tag) {
   if (info.Length() == 0) {
     info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
   }
 
   vector<pair<string, string> > words;
@@ -98,6 +157,7 @@ NAN_METHOD(tag) {
 NAN_METHOD(extract) {
   if (info.Length() != 2) {
     info.GetReturnValue().Set(Nan::New<v8::Boolean>(false));
+    return;
   }
 
   string sentence = *(String::Utf8Value(info[0]->ToString()));