+++ lib/nodejieba.h
@@ -5,6 +5,9 @@
 
 extern NAN_METHOD(load);
 extern NAN_METHOD(cut);
+extern NAN_METHOD(cutAll);
+extern NAN_METHOD(cutHMM);
+extern NAN_METHOD(cutForSearch);
 extern NAN_METHOD(tag);
 extern NAN_METHOD(extract);
 extern NAN_METHOD(insertWord);